Refactoring/routers (#225)
* refactoring/routers: move routers to new folder

* refactoring/routers: move routers to new folder
